refactor(music): Moved playback control commands to new file
