+++ src/commands/music/mod.rs
@@ -1,9 +1,11 @@
 use serenity::framework::standard::macros::group;
 
 mod channels;
+mod play;
 mod queue;
 
 use channels::*;
+use play::*;
 use queue::*;
 
 #[group]
